pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "django-wtf"
version = "0.0.1"
description = "django.wtf"
requires-python = ">=3.12"
authors = [{ name = "Dani Hodovic", email = "dani-hodovic@example.com" }]
dependencies = [
  "psycopg[binary,pool]>=3.3.3",
  "whitenoise>=6.12.0",
  "django-anymail>=14.0",
  "gunicorn>=25.1.0",
  "django-extensions>=4.1",
  "django-environ>=0.13.0",
  "django-model-utils>=5.0.0",
  "django-meta @ git+https://github.com/ivanchenkodmitry/django-meta@patch-1",
  "django-health-check>=4.0.6",
  "django-allauth>=65.14.3",
  "django-filter>=25.2",
  "django-debug-toolbar>=6.2.0",
  "celery>=5.6.2",
  "Django>=6.0.2",
  "django-celery-beat @ git+https://github.com/celery/django-celery-beat.git@main",
  "django-allauth-ui>=1.8.1",
  "django-widget-tweaks>=1.5.1",
  "redis>=7.2.1",
  "requests>=2.32.5",
  "django-redis>=6.0.0",
  "ptipython>=1.0.1",
  "ptpython>=3.0.32",
  "argon2-cffi>=25.1.0",
  "django-constance>=4.3.4",
  "django-toolshed>=0.5.9",
  "django-tailwind>=4.4.2",
  "django-browser-reload>=1.21.0",
  "django-cacheops>=7.2",
  "django-admin-interface>=0.32.0",
  "superrequests>=0.1.2",
  "praw>=7.8.1",
  "hanzidentifier>=1.3.0",
  "django-user-agents>=0.4.0",
  "django-celery-results>=2.6.0",
  "Markdown>=3.10.2",
  "django-watson>=1.6.3",
  "wagtail>=7.3",
  "wagtail-code-blog>=0.6.2",
  "wagtail-markdown>=0.13.0",
  "mattermostdriver>=7.3.2",
  "django-storages[boto3]>=1.14.6",
  "django-custom-error-views>=0.3.1",
  "django-prometheus @ git+https://github.com/django-commons/django-prometheus.git@master",
  "django-waffle>=5.0.0",
  "django-admin-site-search>=1.1.0",
  "django-admin-shellx>=0.3.3",
  "channels-redis>=4.3.0",
  "django-htmx>=1.27.0",
  "pygments>=2.19.2",
  "pypandoc-binary>=1.16.2",
  "django-o11y[all]>=0.5.2",
]

[dependency-groups]
dev = [
  "pudb>=2022.1.3",
  "pytest>=9.0.2",
  "pytest-django>=4.12.0",
  "coverage>=7.13.4",
  "factory-boy>=3.3.3",
  "mypy>=1.19.1",
  "pylint>=4.0.5",
  "pylint-django>=2.7.0",
  "pytest-cov>=7.0.0",
  "django-stubs>=5.2.9",
  "pylint-celery>=0.3",
  "Werkzeug>=3.1.3",
  "django-coverage-plugin>=2.0.4",
  "responses>=0.25.7",
  "types-requests>=2.32.4.20260107",
  "types-python-dateutil>=2.9.0.20250516",
  "types-Markdown>=3.8.0.20250413",
  "djlint>=1.36.4",
  "pytest-factoryboy>=2.7.0",
  "rich>=14.0.0",
  "ruff>=0.15.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
package = false

[tool.pytest.ini_options]
addopts = "--ds=config.settings.test --reuse-db --pdbcls=pudb.debugger:Debugger"
python_files = "tests.py test_*.py"
norecursedirs = ".git */migrations/* */static/* venv .venv .poetry .virtualenv node_modules"

[tool.ruff]
line-length = 88
target-version = "py312"
extend-exclude = ["migrations", "fixtures", "infra"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "DJ"]
ignore = [
  "E501",
  "E402",
  "F401",
  "F403",
  "F405",
  "I001",
  "B010",
  "B904",
  "DJ001",
]

[tool.ruff.lint.isort]
known-first-party = ["django_wtf", "config"]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401"]
"manage.py" = ["F401"]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]
exclude = [".venv", "migrations.*"]

[tool.django-stubs]
django_settings_module = "config.settings.test"

[tool.pylint.'MASTER']
load-plugins = "pylint_django, pylint_celery"
ignore = "manage.py"

[tool.pylint.'FORMAT']
max-line-length = 150

[tool.pylint.'MESSAGES CONTROL']
disable = "missing-docstring,invalid-name,logging-fstring-interpolation,fixme,duplicate-code"

[tool.pylint.'DESIGN']
max-parents = 13

[tool.pylint.'SIMILARITIES']
min-similarity-lines = 4
ignore-comments = "yes"
ignore-docstrings = "yes"
ignore-imports = "yes"

[tool.djlint]
ignore = "H006,H030,H031,D018"
profile = "django"
preserve_blank_lines = true
use_gitignore = true
format_css = true
format_js = true
